Added fix for special characters used in popover.
Fixed an issue where using special characters in a popover would prevent the popover displaying text, for example: using a comma (,) would remove the rest of the text including and after the comma (,). The issue is caused by using the DOMDocument approach to writing valid XML and wouldn't be a problem if a string literal approach was used instead. (http://php.net/manual/en/domdocument.save.php - Special Characters)